test(ui/view_model/providers): Move and update CustomMockCitySearchViewModel provider test
